requirements/base.txt
# Core dependencies for OnCall Agent

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# GCP integrations
google-cloud-logging==3.9.0
google-auth==2.27.0

# AI/LLM
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
openai==1.10.0
anthropic==0.18.0

# Async support
httpx==0.26.0
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
